+++ documentation/changelog.rst
@@ -18,6 +18,14 @@ Bugfixes
 -----------
 
 
+v0.31.1 | March XX, 2026
+============================
+
+Bugfixes
+-----------
+- Add missing field documentation for ``aggregate-power`` and ``state-of-charge`` fields, which can be used to reference a sensor on which to record extra scheduling results [see `PR #2003 <https://www.github.com/FlexMeasures/flexmeasures/pull/2003>`_]
+
+
 v0.31.0 | February 28, 2026
 ============================